Add initial structure
<h1>Rationale:</h1>

Based on a previous attempt (ChessStar) this version of the app starts
to embrace multiple games, the use of a navigation drawer to eliminate
dialogs, and the heavy use of managed fragments.

This commit exists to enable other team members to contribute code,
issues, ideas, etc. with a goal to reach a beta level app comprising
basic chat, oauth2 sign-in, a TicTacToe game.